auth middleware
cleaning up the auth router
