+++ server/kinde.ts
@@ -2,9 +2,13 @@ import {
   createKindeServerClient,
   GrantType,
   type SessionManager,
+  type UserType,
 } from '@kinde-oss/kinde-typescript-sdk';
 import { type Context } from 'hono';
 import { getCookie, setCookie, deleteCookie } from 'hono/cookie';
+
+import { createFactory, createMiddleware } from 'hono/factory';
+
 // Client for authorization code flow
 export const kindeClient = createKindeServerClient(
   GrantType.AUTHORIZATION_CODE,
@@ -51,4 +55,26 @@ export const sessionManager = (c: Context): SessionManager => ({
       deleteCookie(c, key);
     });
   },
+});
+
+type Env = {
+  Variables: {
+    user: UserType;
+  };
+};
+
+export const getUser = createMiddleware<Env>(async (c, next) => {
+  try {
+    const manager = sessionManager(c);
+    const isAuthenticated = await kindeClient.isAuthenticated(manager);
+    if (!isAuthenticated) {
+      return c.json({ error: 'Unauthorized' }, 401);
+    }
+    const user = await kindeClient.getUserProfile(manager);
+    c.set('user', user);
+    await next();
+  } catch (e) {
+    console.error(e);
+    return c.json({ error: 'Unauthorized' }, 401);
+  }
 });
